+++ packages/block-editor/src/hooks/block-bindings.js
