+++ packages/block-editor/src/hooks/index.js
@@ -29,7 +29,7 @@ import childLayout from './layout-child';
 import contentLockUI from './content-lock-ui';
 import './metadata';
 import blockHooks from './block-hooks';
-import blockBindingsPanel from './block-bindings-panel';
+import blockBindingsPanel from './block-bindings';
 import './block-renaming';
 import './use-bindings-attributes';
